backend/requirements.txt
fastapi==0.116.2
uvicorn[standard]== 0.36.0
python-dotenv==1.0.0
pydantic==2.5.0
google-cloud-aiplatform==1.111.0
websockets== 15.0.1
python-multipart==0.0.6
httpx==0.28.1
